- Move functionaly out of main to utils and argparser - Create config, errors, custom_types
