+++ visual_excuses/config.py
@@ -1,0 +1,4 @@
+teampkgs = "http://reqorts.qa.ubuntu.com/reports/m-r-package-team-mapping.json"
+
+people_canonical = "https://people.canonical.com"
+excuses_root_url = people_canonical + "/~ubuntu-archive/proposed-migration/"
