+++ visual_excuses/custom_types.py
@@ -1,0 +1,2 @@
+ExcusesData = dict[str, dict[str, str | int | list[str] | list[int]]]
+PackagesByTeam = dict[str, list[str]]
