+++ visual_excuses/errors.py
@@ -1,0 +1,7 @@
+class VisualExcusesError(Exception):
+    pass
+
+
+class AccessError(VisualExcusesError):
+    # todo: Add the status code in the error message
+    pass
